+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - Fix for missing tags from new messages due to double processing of messages.
 
+### Added
+
+- Added push notification properties to session init optional params
+
 ## [1.10.6] - 2025-01-28
 
 ### Added